Change BinaryAnalyzerContext.MaxFileSizeInKilobytes from int to long to match with the IAnalysisContext interface

+++ src/BinSkim.Sdk/BinaryAnalyzerContext.cs
@@ -135,6 +135,6 @@ namespace Microsoft.CodeAnalysis.IL.Sdk
                 "A shared CompilerDataLogger instance that will be passed to all skimmers.");
 
 
-        public int MaxFileSizeInKilobytes { get; set; }
+        public long MaxFileSizeInKilobytes { get; set; }
     }
 }
